+++ jobbergate-api/jobbergate_api/cli/db.py
@@ -10,7 +10,7 @@ app = typer.Typer()
 
 
 @app.command()
-def login(test: bool = typer.Option(False, help="Log into the test database instead"),):
+def login(test: bool = typer.Option(False, help="Log into the test database instead")):
     """
     Logs into a local database.
     """
@@ -53,3 +53,38 @@ def start_test():
 def start_all():
     start_dev()
     start_test()
+
+
+@app.command()
+def migrate(
+    message: str = typer.Option("Unlabeled migration", help="The message to attach to the migration"),
+    blank: bool = typer.Option(False, help="Produce a blank migration"),
+):
+    """
+    Logs into a local database.
+    """
+    commands = [
+        "alembic",
+        "--config=alembic/alembic.ini",
+        "revision",
+        f"--message={message}",
+    ]
+    if not blank:
+        commands.append("--autogenerate")
+
+    subprocess.run(commands)
+
+
+@app.command()
+def upgrade(target: str = typer.Option("head", help="The migration to which the db should be upgraded"),):
+    """
+    Logs into a local database.
+    """
+    commands = [
+        "alembic",
+        "--config=alembic/alembic.ini",
+        "upgrade",
+        target,
+    ]
+
+    subprocess.run(commands)