+++ jobbergate-cli/tests/subapps/job_scripts/test_app.py
@@ -4,9 +4,14 @@ import shlex
 
 import httpx
 import snick
-import yaml
 
-from jobbergate_cli.schemas import ListResponseEnvelope, Pagination
+from jobbergate_cli.schemas import (
+    ApplicationResponse,
+    JobScriptResponse,
+    JobSubmissionResponse,
+    ListResponseEnvelope,
+    Pagination,
+)
 from jobbergate_cli.subapps.job_scripts.app import (
     HIDDEN_FIELDS,
     JOB_SUBMISSION_HIDDEN_FIELDS,
@@ -76,7 +81,7 @@ def test_get_one__success(
     assert result.exit_code == 0, f"get-one failed: {result.stdout}"
     mocked_render.assert_called_once_with(
         dummy_context,
-        dummy_job_script_data[0],
+        JobScriptResponse(**dummy_job_script_data[0]),
         title="Job Script",
         hidden_fields=HIDDEN_FIELDS,
     )
@@ -91,15 +96,12 @@ def test_create__full_run_including_non_fast_mode_and_job_submission(
     dummy_job_submission_data,
     dummy_domain,
     dummy_render_class,
-    dummy_config_source,
     cli_runner,
     tmp_path,
     attach_persona,
     mocker,
 ):
-
-    application_data = dummy_application_data[0]
-    application_id = application_data["id"]
+    application_response = ApplicationResponse(**dummy_application_data[0])
 
     job_script_data = dummy_job_script_data[0]
     job_script_id = job_script_data["id"]
@@ -117,21 +119,12 @@ def test_create__full_run_including_non_fast_mode_and_job_submission(
     sbatch_params = " ".join(f"--sbatch-params={i}" for i in (1, 2, 3))
 
     param_file_path = tmp_path / "param_file.json"
-    param_file_path.write_text(
-        json.dumps(
-            dict(
-                qux="qux",
-                quux="quux",
-            )
-        )
-    )
+    param_file_path.write_text(json.dumps(dict(foo="oof")))
 
     dummy_render_class.prepared_input = dict(
         foo="FOO",
         bar="BAR",
         baz="BAZ",
-        qux="QUX",
-        quux="QUUX",
     )
 
     attach_persona("dummy@dummy.com")
@@ -140,11 +133,11 @@ def test_create__full_run_including_non_fast_mode_and_job_submission(
     mocked_render = mocker.patch("jobbergate_cli.subapps.job_scripts.app.render_single_result")
     mocked_fetch_application_data = mocker.patch(
         "jobbergate_cli.subapps.job_scripts.app.fetch_application_data",
-        return_value=application_data,
+        return_value=application_response,
     )
     mocked_create_job_submission = mocker.patch(
         "jobbergate_cli.subapps.job_scripts.app.create_job_submission",
-        return_value=job_submission_data,
+        return_value=JobSubmissionResponse(**job_submission_data),
     )
     mocker.patch.object(
         importlib.import_module("inquirer.prompt"),
@@ -157,7 +150,7 @@ def test_create__full_run_including_non_fast_mode_and_job_submission(
             snick.unwrap(
                 f"""
                 create --name=dummy-name
-                       --application-id={application_id}
+                       --application-id={application_response.id}
                        --param-file={param_file_path}
                        {sbatch_params}
                 """
@@ -168,7 +161,7 @@ def test_create__full_run_including_non_fast_mode_and_job_submission(
     assert result.exit_code == 0, f"create failed: {result.stdout}"
     assert mocked_fetch_application_data.called_once_with(
         dummy_context,
-        id=application_id,
+        id=application_response.id,
         identifier=None,
     )
     assert mocked_create_job_submission.called_once_with(
@@ -178,39 +171,38 @@ def test_create__full_run_including_non_fast_mode_and_job_submission(
     )
     assert create_route.called
     content = json.loads(create_route.calls.last.request.content)
-    unpacked_content = {
-        **content,
-        "param_dict": json.loads(content["param_dict"]),
-    }
-    assert unpacked_content == dict(
+    assert content == dict(
         param_dict=dict(
-            **yaml.safe_load(dummy_config_source),
-            job_script_name="dummy-name",
-            application_id=application_id,
-            foo="FOO",
-            bar="BAR",
-            baz="BAZ",
-            qux="qux",
-            quux="quux",
+            application_config=dict(
+                foo="oof",
+                bar="BAR",
+                baz="BAZ",
+            ),
+            jobbergate_config=dict(
+                default_template="test-job-script.py.j2",
+                job_script_name=None,
+                output_directory=".",
+                supporting_files=None,
+                supporting_files_output_name=None,
+                template_files=["test-job-script.py.j2"],
+            ),
         ),
+        application_id=application_response.id,
         job_script_name="dummy-name",
-        sbatch_params_0="1",
-        sbatch_params_1="2",
-        sbatch_params_2="3",
-        sbatch_params_len=3,
+        sbatch_params=["1", "2", "3"],
     )
 
     mocked_render.assert_has_calls(
         [
             mocker.call(
                 dummy_context,
-                job_script_data,
+                JobScriptResponse(**job_script_data),
                 title="Created Job Script",
                 hidden_fields=HIDDEN_FIELDS,
             ),
             mocker.call(
                 dummy_context,
-                job_submission_data,
+                JobSubmissionResponse(**job_submission_data),
                 title="Created Job Submission (Fast Mode)",
                 hidden_fields=JOB_SUBMISSION_HIDDEN_FIELDS,
             ),
@@ -225,15 +217,12 @@ def test_create__with_fast_mode_and_no_job_submission(
     dummy_application_data,
     dummy_job_script_data,
     dummy_domain,
-    dummy_config_source,
     cli_runner,
     tmp_path,
     attach_persona,
     mocker,
 ):
-
-    application_data = dummy_application_data[0]
-    application_id = application_data["id"]
+    application_response = ApplicationResponse(**dummy_application_data[0])
 
     job_script_data = dummy_job_script_data[0]
 
@@ -251,11 +240,9 @@ def test_create__with_fast_mode_and_no_job_submission(
     param_file_path.write_text(
         json.dumps(
             dict(
-                foo="foo",
-                bar="bar",
-                baz="baz",
-                qux="qux",
-                quux="quux",
+                foo="oof",
+                bar="rab",
+                baz="zab",
             )
         )
     )
@@ -266,7 +253,7 @@ def test_create__with_fast_mode_and_no_job_submission(
     mocked_render = mocker.patch("jobbergate_cli.subapps.job_scripts.app.render_single_result")
     mocked_fetch_application_data = mocker.patch(
         "jobbergate_cli.subapps.job_scripts.app.fetch_application_data",
-        return_value=application_data,
+        return_value=application_response,
     )
     result = cli_runner.invoke(
         test_app,
@@ -274,7 +261,7 @@ def test_create__with_fast_mode_and_no_job_submission(
             snick.unwrap(
                 f"""
                 create --name=dummy-name
-                       --application-id={application_id}
+                       --application-id={application_response.id}
                        --param-file={param_file_path}
                        --fast
                        --no-submit
@@ -286,36 +273,35 @@ def test_create__with_fast_mode_and_no_job_submission(
     assert result.exit_code == 0, f"create failed: {result.stdout}"
     assert mocked_fetch_application_data.called_once_with(
         dummy_context,
-        id=application_id,
+        id=application_response.id,
         identifier=None,
     )
     assert create_route.called
     content = json.loads(create_route.calls.last.request.content)
-    unpacked_content = {
-        **content,
-        "param_dict": json.loads(content["param_dict"]),
-    }
-    assert unpacked_content == dict(
+    assert content == dict(
         param_dict=dict(
-            **yaml.safe_load(dummy_config_source),
-            job_script_name="dummy-name",
-            application_id=application_id,
-            foo="foo",
-            bar="bar",
-            baz="baz",
-            qux="qux",
-            quux="quux",
+            application_config=dict(
+                foo="oof",
+                bar="rab",
+                baz="zab",
+            ),
+            jobbergate_config=dict(
+                default_template="test-job-script.py.j2",
+                job_script_name=None,
+                output_directory=".",
+                supporting_files=None,
+                supporting_files_output_name=None,
+                template_files=["test-job-script.py.j2"],
+            ),
         ),
+        application_id=application_response.id,
         job_script_name="dummy-name",
-        sbatch_params_0="1",
-        sbatch_params_1="2",
-        sbatch_params_2="3",
-        sbatch_params_len=3,
+        sbatch_params=["1", "2", "3"],
     )
 
     mocked_render.assert_called_once_with(
         dummy_context,
-        job_script_data,
+        JobScriptResponse(**job_script_data),
         title="Created Job Script",
         hidden_fields=HIDDEN_FIELDS,
     )
@@ -357,7 +343,7 @@ def test_update__makes_request_and_renders_results(
     assert result.exit_code == 0, f"update failed: {result.stdout}"
     mocked_render.assert_called_once_with(
         dummy_context,
-        new_job_script_data,
+        JobScriptResponse(**new_job_script_data),
         title="Updated Job Script",
         hidden_fields=HIDDEN_FIELDS,
     )